+++ jenkins/CHANGELOG.md
@@ -1,3 +1,21 @@
+0.0.14
+
+* Version bump to rebuild
+
+---
+
+0.0.13
+
+* Version bump to diagnose and fix freebsd-13 build error
+
+---
+
+0.0.12
+
+* Rebuild for latest versions
+
+---
+
 0.0.11
 
 * Adjustments to host key management for ssh