+++ nomad-server/CHANGELOG.md
@@ -1,3 +1,11 @@
+2.0.14
+
+* Fix /var/tmp/nomad issue
+* Add parameter to set regions
+
+---
+
+
 2.0.13
 
 * Automatic job import, plan, from copied-in job files in .nomad format